* Followed Ingo's suggestion, and retricted sending SIGPIPE only to userland   API clients.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@25246 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/network/protocols/tcp/TCPEndpoint.cpp
@@ -723,7 +723,9 @@ TCPEndpoint::SendData(net_buffer *buffer)
 	if (fState == FINISH_SENT || fState == FINISH_ACKNOWLEDGED
 		|| fState == CLOSING || fState == WAIT_FOR_FINISH_ACKNOWLEDGE
 		|| fState == TIME_WAIT) {
-		send_signal(find_thread(NULL), SIGPIPE);
+		// we only send signals when called from userland
+		if (gStackModule->is_syscall)
+			send_signal(find_thread(NULL), SIGPIPE);
 		return EPIPE;
 	}
